+++ test/integration/013_context_var_tests/models/context.sql
@@ -27,4 +27,6 @@ select
     '{{ run_started_at }}' as run_started_at,
     '{{ invocation_id }}'  as invocation_id,
 
-    '{{ env_var("DBT_TEST_013_ENV_VAR") }}' as env_var
+    '{{ env_var("DBT_TEST_013_ENV_VAR") }}' as env_var,
+    '{{ env_var("DBT_ENV_SECRET_013_SECRET") }}' as env_var_secret,
+    '{{ env_var("DBT_TEST_013_NOT_SECRET") }}' as env_var_not_secret